code/models/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cloudml-hypertune==0.1.0.dev6
cloudpickle==2.0.0
cycler==0.11.0
dask==2021.12.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distributed==2021.12.0
entrypoints==0.3
featuretools==1.3.0
FLAML==0.9.1
flatbuffers==2.0
fonttools==4.28.3
frozenlist==1.2.0
fsspec==2021.11.1
gast==0.4.0
gcsfs==2021.11.1
google-api-core==2.3.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-bigquery==2.31.0
google-cloud-bigquery-storage==2.10.1
google-cloud-core==2.2.1
google-cloud-logging==2.7.0
google-cloud-pubsub==2.9.0
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
h5py==3.6.0
HeapDict==1.0.1
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
imbalanced-learn==0.8.1
imblearn==0.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.3.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
libcst==0.3.23
lightgbm==3.3.1
llvmlite==0.38.0
locket==0.2.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
missingno==0.5.0
mistune==0.8.4
msgpack==1.0.3
multidict==5.2.0
multimethod==1.6
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.10
numba==0.55.0
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandas-profiling==3.1.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
patsy==0.5.2
pexpect==4.8.0
phik==0.12.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scikit-optimize==0.9.0
scikit-plot==0.3.7
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
sortedcontainers==2.4.0
statsmodels==0.13.1
tangled-up-in-unicode==0.1.0
tblib==1.7.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.7
visions==0.7.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
woodwork==0.10.0
wrapt==1.13.3
xgboost==1.3.3
yarl==1.7.2
zict==2.0.0
zipp==3.6.0